requirements.txt
# 恋爱军师.skill 依赖

# 聊天记录解析
# 无额外依赖，使用标准库

# 照片EXIF提取（可选）
Pillow>=9.0.0

# 数据处理
# 标准库已足够